Fix: removed wrong assertion in ESRestTestFeatureService#clusterHasFeature
The check is currently broken - feature check always falls back to historical features, and only then if the feature is not one of the historical ones, and exception (assert) is risen.
This is wrong: it could be that we are testing for a non-historical feature, and the cluster simply does not have it.

The check could be fixed with https://github.com/elastic/elasticsearch/pull/104257, but we decided it has little advantage over the complexity it brings - better remove it.